+++ src/main/java/org/opensearch/knn/index/query/KNNWeight.java
@@ -23,22 +23,26 @@ import org.apache.lucene.util.Bits;
 import org.apache.lucene.util.FixedBitSet;
 import org.opensearch.common.io.PathUtils;
 import org.opensearch.common.lucene.Lucene;
-import org.opensearch.knn.common.FieldInfoExtractor;
 import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.index.KNNSettings;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.VectorDataType;
-import org.opensearch.knn.index.engine.qframe.QuantizationConfig;
 import org.opensearch.knn.index.memory.NativeMemoryAllocation;
 import org.opensearch.knn.index.memory.NativeMemoryCacheManager;
 import org.opensearch.knn.index.memory.NativeMemoryEntryContext;
 import org.opensearch.knn.index.memory.NativeMemoryLoadStrategy;
 import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.index.quantizationService.QuantizationService;
 import org.opensearch.knn.indices.ModelDao;
 import org.opensearch.knn.indices.ModelMetadata;
 import org.opensearch.knn.indices.ModelUtil;
 import org.opensearch.knn.jni.JNIService;
 import org.opensearch.knn.plugin.stats.KNNCounter;
+import org.opensearch.knn.quantization.models.quantizationOutput.QuantizationOutput;
+import org.opensearch.knn.quantization.models.quantizationParams.QuantizationParams;
+import org.opensearch.knn.quantization.models.quantizationState.QuantizationState;
+import org.opensearch.knn.quantization.models.quantizationState.QuantizationStateCacheManager;
+import org.opensearch.knn.quantization.models.quantizationState.QuantizationStateReadConfig;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -72,6 +76,7 @@ public class KNNWeight extends Weight {
     private final ExactSearcher exactSearcher;
 
     private static ExactSearcher DEFAULT_EXACT_SEARCHER;
+    private final QuantizationService quantizationService = QuantizationService.getInstance();
 
     public KNNWeight(KNNQuery query, float boost) {
         super(query);
@@ -227,9 +232,6 @@ public class KNNWeight extends Weight {
             return null;
         }
 
-        // TODO: Use this to get quantization config
-        QuantizationConfig quantizationConfig = FieldInfoExtractor.extractQuantizationConfig(fieldInfo);
-
         KNNEngine knnEngine;
         SpaceType spaceType;
         VectorDataType vectorDataType;
@@ -254,6 +256,25 @@ public class KNNWeight extends Weight {
             vectorDataType = VectorDataType.get(
                 fieldInfo.attributes().getOrDefault(VECTOR_DATA_TYPE_FIELD, VectorDataType.FLOAT.getValue())
             );
+        }
+
+        QuantizationParams quantizationParams = quantizationService.getQuantizationParams(fieldInfo);
+
+        byte[] quantizedVector = null;
+
+        if (quantizationParams != null) {
+            QuantizationState quantizationState = QuantizationStateCacheManager.getInstance()
+                .getQuantizationState(
+                    new QuantizationStateReadConfig(
+                        reader.directory(),
+                        reader.getSegmentName(),
+                        Long.toString(reader.getSegmentInfo().getFieldInfosGen(), Character.MAX_RADIX),
+                        fieldInfo,
+                        quantizationParams.getTypeIdentifier()
+                    )
+                );
+            QuantizationOutput quantizationOutput = quantizationService.createQuantizationOutput(quantizationParams);
+            quantizedVector = (byte[]) quantizationService.quantize(quantizationState, knnQuery.getQueryVector(), quantizationOutput);
         }
 
         List<String> engineFiles = getEngineFiles(reader, knnEngine.getExtension());
@@ -296,10 +317,11 @@ public class KNNWeight extends Weight {
             }
             int[] parentIds = getParentIdsArray(context);
             if (k > 0) {
-                if (knnQuery.getVectorDataType() == VectorDataType.BINARY) {
+                if (knnQuery.getVectorDataType() == VectorDataType.BINARY
+                    || quantizationParams != null && quantizationService.getVectorDataTypeForTransfer(fieldInfo) == VectorDataType.BINARY) {
                     results = JNIService.queryBinaryIndex(
                         indexAllocation.getMemoryAddress(),
-                        knnQuery.getByteQueryVector(),
+                        quantizationParams == null ? knnQuery.getByteQueryVector() : quantizedVector,
                         k,
                         knnQuery.getMethodParameters(),
                         knnEngine,